Do not disregard the NSError field when parsing strings into numbers using the NSNumberFormatter.
Previously, if we passed a string with just a space i.e. " " to the parse method, it would return the number 0 instead of nil.

Thats because we disregard the error field and coerce doubleValue on a nil result, which returns 0.
	Change on 2015/04/06 by shanbhag <shanbhag@google.com>
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=90448047