Provide more system properties, such as os.arch
This brings the system properties provided by jre_emul in line with
those by the Android API. The property "os.arch" is obtained during
the compile time from a macro set in fat_lib.mk. "os.version" attempts
to use the new [NSProcessInfo processInfo].operatingSystemVersion API
first before falling back to other means. The drawback is that if the
x86_64 part of the library is not compiled against OS X 10.10 SDK, the
code will have to use the operatingSystemVersionString method in
NSProcessInfo when used in an OS X app.