Small fix for capturing lambda expressions, dump of basic implementation of AbstractRegressionTest allowing us to borrow tests from Eclipse.
This implementation is incomplete and more than a little broken.

Some of the tests target a different version of Eclipse than our installation, so they break due to the original Eclipse bug.  Some tests use special Eclipse flags to process data a certain way, eg to force an ExpressionMethodReference to be parsed as a TypeMethodReference.  These are largely broken as well.

There are also tests which are designed to fail, to test proper error response within the Eclipse compiler.  These are not useful to us, as we are allowing Eclipse to take care of the parsing duties for us, and to verify code before building an AST.

There are also tests that target specific compiler options, which we may be able to integrate, but which I have not.
	Change on 2015/08/12 by kirbs <kirbs@google.com>
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=100485103