Update Number.java
Hi,
Sorry in advance for my poor english.
I've an error when I translate a class at who I pass a "Long" type when it expect a generic type serializable and comparable:
error: SimpleAutoIndexedEntity.java:15: Bound mismatch: The type Long is not a valid substitute for the bounded parameter <ID extends Serializable & Comparable<ID>> of the type SimpleEntity<ID>
It seems that it was because the Number.java (who is present in the jre_emul.jar file) doesn't implement java.io.Serializable. That's why "Long" class who inherit of the "Number" class can't be serializable.
I've compared to the official Number.java file from the java.lang package and there is other differences on byteValue and shortValue methods.
I don't know if it's necessary, or obligatory, to update this methods too, but this update seems work with it.
Do you think it's good?