Don't redeclare generic private method in subclass
If a superclass contains a method that returns a parameterized type, a subclass shouldn't include a declaration of that method in the .m file if it is a private method since the subclass is restricted from calling that method. In practice, this was causing compilation errors if the method argument list contained types that are not imported in the subclass.

This patch fixes the issue and includes a test case that fails prior to this change.