Provide URLConnection.getErrorStream() (#834)
When the HTTP status >= HTTP_BAD_REQUEST, the session will still proceed so as
to provide the error stream. The error stream is actually just an alias to
responseBodyStream, but because getInputStream() throws an exception when the
HTTP status is an error, this API contract guarantees that responseBodyStream
can only mean one thing (either the result of getInputStream() or that of
getErrorStream()) at any given time.

	Change on 2017/02/10 by lukhnos <lukhnos@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=147196801