Fixed TypeUtil.getReferenceName() for constructors of top-level classes
DeadCodeEliminator uses this method to get method names. The comment of this method is correct. It says "For constructors of top-level classes, this is the name of the class". ProGuard uses class names for constructors. ElementUtil.getName() returns "<init>" for constructors.

	Change on 2017/03/30 by zgao <zgao@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=151715014