Make Reader have @Weak reference to lock
I believe there's a retain cycle leak in Reader.lock. I'm pretty sure this is the case when `this` is used as the lock, since a strong reference to itself is always a retain cycle unless broken elsewhere. In the case of StreamDecoder as used by InputStreamReader, the InputStreamReader instance is used as the StreamDecoder's lock, while the InputStreamReader retains a strong reference to the StreamDecoder resulting in a retain cycle.

I'm not 100% sure if @Weak is safe to use here as I haven't audited every use of the lock constructor parameter and field.