a) generateKeyPair Our requirements is to genereate KeyPairs for several users. I noticed that the created KeyPairs created had all the same identical keys. To solve this problem, in method generateKeyPair, I first deleted any existing public/private keys. Then the keys were always created new. see Class IosRSAKeyPairGenerator.java
b)RSA Sign padding
When RSASignature nativeEngineSign signs the message always with
"kSecPaddingNone" but the nativeEngineVerify method use PKCS1.
The output of the "sign" message is not comapatile with PKCS1. So I
changed RSASignature so that the method nativeEngineSign will also use
PKCS1 padding.
See class IosRSASignature.java

c) engineGeneratePrivate
The KeySpec for PrivateKey generation is PKCS8EncodedKeySpec. The method
engineGeneratePrivate was only considering X509EncodedKeySpec.
To solve this I implemented the check for PKCS8EncodedKeySpec.
See class IosRSAKeyFactory.java

d) endcoding the PrivateKey
The constructor IosRSAPrivateKey (byte [] encoded) was calling
"this(decode(encoded))" and this method tries to load a PrivateKey from
"SecCertificateCreateWithData".
But you cannot laod a Private Key from a DER certificate.
To solve the problem I implemented the method  "iosSecKey =
createPrivatSecKeyRef(encoded);"

e) Note about Public Key format
The iOS PublicKey format is not a PKCS1 standard, it is an ASN.1
sequence encoding of the modules and exponent. If you want to send the
ios PublicKey to for example a JAVA server you have to convert the
PubkeyKey encoding to PKCS1.
To make it a PKCS1 you just need to add the PKCS1 header.
For example the pkcs1 header for 2048 key is :
static byte[] pcks1Header = new byte[] { (byte) 0x30, (byte) 0x82,
(byte) 0x01, (byte) 0x22, (byte) 0x30,
            (byte) 0x0d, (byte) 0x06, (byte) 0x09, (byte) 0x2a, (byte)
0x86, (byte) 0x48, (byte) 0x86, (byte) 0xf7,
            (byte) 0x0d, (byte) 0x01, (byte) 0x01, (byte) 0x01, (byte)
0x05, (byte) 0x00, (byte) 0x03, (byte) 0x82,
            (byte) 0x01, (byte) 0x0f, (byte) 0x00 };