Add support for Character unboxing on cast
Casting java.lang.Character to a primitive currently attempts to cast it directly into its boxed type (as is done with Object types), which will fail in runtime and is inconsistent with java.

According to the JLS:
"Casting contexts allow the use of ... an unboxing conversion optionally followed by a widening primitive conversion"

Character is the only source for such Unboxing+widening conversions that is not a Number, and it converts to int, long, float and double.

This commit adds a special casing for Character, on top of the existning case for Number conversions.