Project health: build guava with --swift-friendly.
The javac AST returned by the JDK-1.8 does not contain type annotations present in type-variable parameters. For example, @Nullable is not present in the AST corresponding to:

void notifyNewValue(@Nullable T newValue)

If this code lives in a package annotated with @ParametersAreNonnullByDefault, then the generated code will be annotated with non-null.

- (void)notifyNewValueWithId:(id __nonnull)newValue;

This is causing the following type of compilation errors:

error: null passed to a callee that requires a non-null argument [-Werror,-Wnonnull]
    [valueReference notifyNewValueWithId:nil];

The simplest workaround is to ignore some of the package-info files when transpiling.

This is fixed in JDK 9+.

	Change on 2019/06/18 by antoniocortes <antoniocortes@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=253784644