Automated g4 rollback of changelist 360297507.
*** Reason for rollback ***

Generated code passes two arguments to a one-argument function. TAP tests didn't catch this because it only shows up with --j2objc_translation_flags=-use-arc, which no projects are currently using.

*** Original change description ***

When weak fields are initialized with constructor invocations in ARC, retain/autorelease before assignment. ARC releases the object but doesn't retain it with a weak field, causing it to immediate deallocate.

PiperOrigin-RevId: 360440225