Remove STATIC_LIBRARY build dependenxy reference
The building of the framework has a build dependency to "$(BUILD_DIR)/lib$(STATIC_LIBRARY_NAME).a"
This the binary of the iphone platform. This causes, that iphone binary must be build and added
to the xcframework, even when not wantend.

That the required libs build before the xcframework is built, is already satisfied
with the "dist" dependency.