Fix xcarchive platform slices
There was a bug in the construction of the fat library paths for generating the xcframework.
That caused, that some archtitecture weren't added or caused an build error "Both xx and yyy
represent two equivalent library definitions".

For each platform the fat library is now added, when at least one archictecutre of the platform
is added to environment variable J2OBJC_ARCHS.