Rename android.os to android.internal.
J2obcj internal AsyncTask or SystemClock are included in j2objc for technical reasons and shouldn't be used in client code (and were never intended to). This move makes this explicit, avoiding accidental future dependencies. AsyncTask is deprecated and SystemClock is largely redundant with methods contained in "standard" Java libraries.

PiperOrigin-RevId: 410844558