Update headers within jre_emul and J2ObjC's proto runtime to be compatible with explicit modules for Swift
`#pragma clang system_header` is removed as these are not system headers and it is incompatible with explicit module compilation, resulting in `error: #pragma system_header ignored in main file [-Werror,-Wpragma-system-header-outside-header]`

Missing `#include`s are added to address errors when compiling the headers independently.

A cast is added within `IOSReflection.h` to address an incompatible assignment (`error: cannot initialize a variable of type 'const char *' with an rvalue of type 'const void *'`).

Parameters named `this` are renamed as `this` is a keyword when the header is parsed as C++ (`error: invalid parameter name: 'this' is a keyword`).

PiperOrigin-RevId: 415623605