Undo performance optimization around standard data format.
Eagerly creating the standard date format, and then re-using as much as possible, is purely a performance optimization that can possibly lead to issues due to the stateful-ness of DateFormat.

In this cl we remove this performance optimization to simplify the code and avoid any possible state issues in DateFormat by creating a new parser on every invocation. This is similar to how the fallback formats are already handled anyway.

PiperOrigin-RevId: 576304654