Generate private class declaration nullability annotations for `-Xnull-marked`.
When `-Xnull-marked` is on, packages with the `@NullMarked` annotation have their interface declarations generated within Objective-C nullability audited regions (which make otherwise unannotated pointers `nonnull`).  This includes inner classes from Java which are generated within the Objective-C implementation files of the outer classes in which they are contained in Java.  This change adds the generation of the per-Element nullability annotations for those private classes' interfaces so they are both internally-consistent with the `nonnull` annotations from `@NullMarked` and that private subclasses' interfaces remain consistently annotated with the public interfaces they override.

PiperOrigin-RevId: 580949777