Improved Interfaces for Swift
Added a new annotation named `@SwiftName` and has one argument for the name. It will be available for package, class/interface, and methods. Also added a build flag named `--swift-naming` and it has the same effect as the annotation.

When the annotation is added on the package it use the java name as the swift name for all classes and methods. When added on a class or method level, it will allow renaming of that class or method for Swift. If no argument is added it use the Java name.

Currently enums make a NS_ENUM and Class type. The annotation will fold the class type into the enum type in Swift. There is also two new methods for converting between the types.

PiperOrigin-RevId: 609123126