Automated g4 rollback of changelist 638827797.
*** Reason for rollback ***

Roll-forward: previous static reference to java.lang.reflect.GenericDeclaration in Java failed because it's an interface, and so was transpiled to an ObjC protocol. This CL makes a native static reference to the protocol's backing class.

*** Original change description ***

Automated g4 rollback of changelist 638722935.

*** Reason for rollback ***

com.google.gson.internal.reflect.ReflectionHelper$RecordSupportedHelper was unable to reflect on java.lang.reflect.GenericDeclaration.

*** Original change description ***

Supports linking the core jre_emul subset library without the `-ObjC` flag, by making static references to classes that are dynamically loaded. The static references ensure that the classes continue to be linked into the app.

PiperOrigin-RevId: 639112548