Update InputStreamReader and OutputStreamReader to use their lock instead of themselves as the encoder/decoder lock.
Stream decoder will continue to take in a lock via the public factory methods. The internal initializers will still call super with the lock. To ensure the initializers will not be called, they will be made private.

It’s up to the InputStreamReader to pass in the correct lock instead of itself. So to fix this, the lock created by the InputStreamReader’s superclass (Reader) at construction time will be passed in instead.

Did the same thing for the OutputStreamWriter. (StreamEncoder constructors were already private)

This was causing a retain cycle in iOS.

Additionally, the reader and decoder were using different objects as their lock which may lead to some contention

PiperOrigin-RevId: 678332165