Detect include cycles when segmented headers are disabled.
Some java libraries require segmented headers because they otherwise result in #include cycles, but the segmented headers are not modular. A library without include cycles can be transpiled without segmented headers and have modular headers, but can start failing at a later time if the java code is modified to add a cycle. This change adds cycle detection for translation operations without segmented headers, and fails with a notice to enable segmented headers if a cycle is detected.

PiperOrigin-RevId: 758306647