Setting Null-Reference on for originalBufferObject on MemoryRef
In the generated code, this results in a cyclic dependency. Since the mechanism was created to hold a reference to the original Buffer, after the "live" reference might has been deallocated, this seems not to be a working mechanism in Objective-C.

This change sets the originalBufferObject to always null (like on the other constructor already) to avoid this behavior. Will need to test for side-effects