Remove `modifiers` for `VariableDeclarationStatements`.
The Java source modifiers are already kept in the associated `VariableElement` and it is never used from the statement except for the `DebugASTPrinter` (only used to visually inspect the source).

This is a cleanup because a follow-up will introduce objective-c modifiers to all three flavors of variable declarations in j2objc (`VariableDeclarationStatement`, `VariableDeclarationExpression` and `SingleVariableDeclaration`). Objective-C `__block` modifier is necessary to allow variables be mutated in the newly introduced block expressions.

PiperOrigin-RevId: 827569524